NotFrosh Registration
Added NotFrosh registration to the frosh page. Ticket sales are done
through stripe.

There is a file called config.json in the data folder that can be used
to configure the email address that sends the confirmation. The config
file also allows for a stripe account to be set up.

Some things that may be nice to add:
-Redirect to a receipt page